feat(infra): Initial Infra-as-Code with AWS and Pulumi
closes #11

This commit does a lot so let's try and document what is going on.

We're using Pulumi.com as our Infra-as-Code solution.
It is similar to Terraform but as code instead of configuration files.
The two main benefits are:
1. More familiar syntax (js code vs yaml configs)
2. Auto-complete (typescript)

There are two Github Actions workflows being set:
1. On Pull Request — Does a dry-run of the pulumi script
2. On Push / Merge to master — Takes changes to infra to production

Pulumi.planx.yml holds configuration and secrets.
The encryption key is hold by Pulumi's servers.
We could've have used only Github Secrets
but this approach has two benefits:
1. Easier to manage multiple environments
2. Easier to get secrets on local machine (pulumi config get)

The `docker` CLI doesn't accept Dockerfiles with a custom name.
It only expects `directory/Dockerfile`.
So we've renamed
- `*/Dockerfile` -> `*/Dockerfile.dev`
- `*/Dockerfile.prod` -> `*/Dockerfile`
and updated docker-compose.yml accordingly.

Locally, docker-compose allows us to attach volumes
from the host filesystem into the container.
On AWS that's not possible.
So we're copying migrations and metadata into the docker image.

AWS doesn't have a Postgis engine.
The recommended way is to execute `CREATE EXTENSION postgis`
on any Postgres instances.
In order to do that, this commit adds to an existing migration file.

The Pulumi script requires access to Cloudflare's API in order to
set DNS records for:
1. Verifying ownership of the domain so AWS will let API Gateway use it
2. Pointing the domain to API Gateway

EOF