feat: create new session
Closes #36

I was in two minds about where to store and retrieve
the flow id, data, and version.

On the one hand,
we already have a store in place,
that we could query for the id and data,
and that we could easily extend to add the version field.

On the other hand,
that store really belongs to the editor,
not to the preview.

I've decided to go with the latter.

I hope this will further decouple the preview pane from the editor,
and will make things simpler in the public interface.

This is my first contribution to the editor package,
so I might be missing something.