feat: flows.data_merged
Adds new computed field to the "flows" GraphQL Object.
It's denormalized flow data.
It merges portals' flows into the main flow (recursively).

As it regards the order of the edges in the array,
this algorithm always appends to the end of the array of edges,
to keep things simple.

This works for simpler cases
but I have a gut feeling some things might be out of order
in flows with multiple portals positioned in specific places.