fix: allow public to add session_events.parent_node_id
This was a bug,
that was failing hasura/tests.

We explicitly specify which columns
public is allowed to INSERT to
because it'd be weird and insecure
to allow a public user
to set the created_at, updated_at, and id fields
of a session_events row.

But when we added session_events.parent_node_id,
we forgot to update permissions.

Hence the need to integrate hasura/tests
into our CD/CI pipeline.