refact(ICONS): use types to enforce exhaustive icons for node types
While this commit still allows us
to leave some node types without an icon,
it enforces that we make that explicit
by setting it to undefined.

Which means when we add a new node type
typescript will prompt us to choose an icon
or consciously decide not to add one.