feat: add FindPropertyMerged
If you think this commit does "nothing" then I've succeeded in my goal.

This is an intermediate step towards a better, unified FindProperty.

This component is a merge of two existing components,
FindProperty and PropertyInformation.

In the next PR I will be removing the deprecated ones
and will rename FindPropertyMerged into FindProperty.

https://trello.com/c/1NyEs2G4/781-merge-find-property-and-property-information-components-into-a-single-component-called-find-property