break: remove metabase
Metabase still isn't working, as we're getting 504 Gateway Time-out
errors. That being said, John just noticed that Metabase created a large
number of tables in our Postgres database - more than we're comfortable
with. So we're going to move Metabase to a separate schema/database.
For now, the plan is to (1) remove metabase (this commit) and (2) revert
staging to before metabase had been introduced. Luckily this
Metabae never reached production so we don't have to do it in production.