fix: keep squares when resizing window
`minHeight` expects a string, so when we pass
`buttonRef.current?.offsetWidth` that's actually a value not a reference.
In order to update it whenever the window size is updated,
and keep the square a square,
we need to use a hook to watch window resizes.

(cherry picked from commit 4b4a9eccd2aa9cf6245cbf5fb368da8bdf246d54)