Lock planx-document-templates to a fixed version
We normally update packages explicitly and I think that should be the
case for planx-document-templates too. It does add a little bit of
additional work but it means that updates have a chance to be
sense-checked and released explicitly.

The ideal solution for managing multiple interdependent packages would
be to use some monorepo tooling. Maybe its worth exploring those instead
of some of these less-than-ideal workflows (or we just put up with the
wonky workflow for the time-being).

If we do want to just track main for this pacakge, do we also need a
testing workflow where we link the module in one of the consumers,
install and build to avoid directly breaking main in
planx-document-templates?.

Are there other solutions to avoiding issues and laborious workflows
here?